refactor part 1
